# web/scripts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "goldilocks-core",
    "version": "0.1.0"
  },
  "paths": {
...
    "/structure/load": {
      "post": {
        "summary": "Load Structure",
        "description": "Validate inline structure content and return a Structure Document.",
        "operationId": "load_structure_structure_load_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructureSource"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructureDocumentModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Tasks",
        "description": "Describe every registered Core task with stable identifiers.",
        "operationId": "tasks_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCatalogueModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/recommend": {
      "post": {
        "summary": "Recommend",
        "description": "Run the recommend preset.",
        "operationId": "recommend_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/generate": {
      "post": {
        "summary": "Generate",
        "description": "Run the generate preset and return generated input contents.",
        "operationId": "generate_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/compute": {
      "post": {
        "summary": "Compute",
        "description": "Compute only the requested record types.",
        "operationId": "compute_compute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordSetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdviceModel": {
        "properties": {
          "smearing": {
            "$ref": "#/components/schemas/SmearingAdviceModel"
          },
          "magnetism": {
            "$ref": "#/components/schemas/MagnetismAdviceModel"
          },
          "spin_orbit": {
            "$ref": "#/components/schemas/SpinOrbitAdviceModel"
          },
          "pseudopotentials": {
            "$ref": "#/components/schemas/PseudopotentialAdviceModel"
          },
          "convergence": {
            "$ref": "#/components/schemas/ConvergenceAdviceModel"
          },
          "vdw": {
            "$ref": "#/components/schemas/VdwAdviceModel"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "smearing",
          "magnetism",
          "spin_orbit",
          "pseudopotentials",
          "convergence",
          "vdw"
        ],
        "title": "AdviceModel",
        "description": "Serialized ParameterAdvice."
      },
...
      "ComputationRequest": {
        "properties": {
          "structure": {
            "$ref": "#/components/schemas/StructureSource"
          },
          "intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Intent"
              },
              {
                "type": "null"
              }
            ]
          },
          "hints": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Hints"
              },
              {
                "type": "null"
              }
            ]
          },
          "mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "recommend",
                  "generate"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "pseudo_metadata": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PseudoMetadata"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pseudo Metadata"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "structure"
        ],
        "title": "ComputationRequest",
        "description": "Request for the recommend or generate preset.\n\nThe Workbench sends inline structure content and identifies pseudos by\nfilename/library; it never supplies server paths. ``output_dir`` and\n``pseudo_root`` are rejected at the HTTP boundary."
      },
...
      "CoreResultResponse": {
        "properties": {
          "core_version": {
            "type": "string",
            "title": "Core Version"
          },
          "intent": {
            "$ref": "#/components/schemas/IntentModel"
          },
          "analysis": {
            "$ref": "#/components/schemas/AnalysisModel"
          },
          "advice": {
            "$ref": "#/components/schemas/AdviceModel"
          },
          "k_points": {
            "$ref": "#/components/schemas/KPointSelectionModel"
          },
          "selection": {
            "$ref": "#/components/schemas/SelectionModel"
          },
          "generated_files": {
            "items": {
              "$ref": "#/components/schemas/GeneratedFileModel"
            },
            "type": "array",
            "title": "Generated Files"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "bundle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BundleModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "core_version",
          "intent",
          "analysis",
          "advice",
          "k_points",
          "selection"
        ],
        "title": "CoreResultResponse",
        "description": "Response of the recommend and generate presets."
      },
      "ErrorDetail": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "kind",
          "message",
          "status"
        ],
        "title": "ErrorDetail",
        "description": "One structured transport failure."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Structured failure envelope returned by the transport."
      },
...
      "Intent": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "quantum_espresso"
          },
          "task": {
            "type": "string",
            "title": "Task",
            "default": "scf_single_point"
          },
          "functional": {
            "type": "string",
            "title": "Functional",
            "default": "PBEsol"
          },
          "pseudo_mode": {
            "type": "string",
            "title": "Pseudo Mode",
            "default": "efficiency"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "Intent",
        "description": "Calculation intent fields."
      },
      "IntentModel": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "task": {
            "type": "string",
            "title": "Task"
          },
          "functional": {
            "type": "string",
            "title": "Functional"
          },
          "pseudo_mode": {
            "type": "string",
            "title": "Pseudo Mode"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "code",
          "task",
          "functional",
          "pseudo_mode"
        ],
        "title": "IntentModel",
        "description": "Serialized CalculationIntent."
      },
...
      "RecordQuery": {
        "properties": {
          "structure": {
            "$ref": "#/components/schemas/StructureSource"
          },
          "outputs": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "analysis",
                    "advice",
                    "k_points",
                    "selection",
                    "generated_files"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputs"
          },
          "intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Intent"
              },
              {
                "type": "null"
              }
            ]
          },
          "hints": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Hints"
              },
              {
                "type": "null"
              }
            ]
          },
          "pseudo_metadata": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PseudoMetadata"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pseudo Metadata"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "structure"
        ],
        "title": "RecordQuery",
        "description": "Request selecting Core records to compute.\n\n``outputs`` is structurally optional so the shared parser can reject a\nmissing selection with a structured failure."
      },
      "RecordSetResponse": {
        "properties": {
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "advice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdviceModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "k_points": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KPointSelectionModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectionModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "generated_files": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/GeneratedFileModel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Files"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "RecordSetResponse",
        "description": "Response of the compute operation, keyed by stable record ids."
      },
...
      "StructureDocumentModel": {
        "properties": {
          "formula": {
            "type": "string",
            "title": "Formula"
          },
          "reduced_formula": {
            "type": "string",
            "title": "Reduced Formula"
          },
          "lattice": {
            "$ref": "#/components/schemas/StructureLatticeModel"
          },
          "sites": {
            "items": {
              "$ref": "#/components/schemas/StructureSiteModel"
            },
            "type": "array",
            "title": "Sites"
          },
          "charge": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Charge"
          },
          "source": {
            "$ref": "#/components/schemas/StructureSourceInfoModel"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "formula",
          "reduced_formula",
          "lattice",
          "sites",
          "source"
        ],
        "title": "StructureDocumentModel",
        "description": "Canonical, transport-safe representation of a parsed structure."
      },
      "StructureLatticeModel": {
        "properties": {
          "matrix": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Matrix"
          },
          "a": {
            "type": "number",
            "title": "A"
          },
          "b": {
            "type": "number",
            "title": "B"
          },
          "c": {
            "type": "number",
            "title": "C"
          },
          "alpha": {
            "type": "number",
            "title": "Alpha"
          },
          "beta": {
            "type": "number",
            "title": "Beta"
          },
          "gamma": {
            "type": "number",
            "title": "Gamma"
          },
          "volume": {
            "type": "number",
            "title": "Volume"
          },
          "pbc": {
            "items": {
              "type": "boolean"
            },
            "type": "array",
            "title": "Pbc"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "matrix",
          "a",
          "b",
          "c",
          "alpha",
          "beta",
          "gamma",
          "volume",
          "pbc"
        ],
        "title": "StructureLatticeModel",
        "description": "Canonical lattice vectors, parameters, and periodicity."
      },
      "StructureSiteModel": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "species": {
            "items": {
              "$ref": "#/components/schemas/StructureSpeciesModel"
            },
            "type": "array",
            "title": "Species"
          },
          "abc": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Abc"
          },
          "xyz": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Xyz"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "label",
          "species",
          "abc",
          "xyz"
        ],
        "title": "StructureSiteModel",
        "description": "One crystallographic site in a canonical Structure Document."
      },
      "StructureSource": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "format": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "cif",
                  "poscar"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content"
        ],
        "title": "StructureSource",
        "description": "Inline structure content; never a client server path."
      },
      "StructureSourceInfoModel": {
        "properties": {
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "inline"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "StructureSourceInfoModel",
        "description": "Origin metadata for a parsed structure."
      },
...
      "TaskCatalogueModel": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskGraphDescriptionModel"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "tasks"
        ],
        "title": "TaskCatalogueModel",
        "description": "Catalogue of every registered task description."
      },
      "TaskGraphDescriptionModel": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "revision": {
            "type": "string",
            "title": "Revision"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageDescriptionModel"
            },
            "type": "array",
            "title": "Stages"
          },
          "presets": {
            "items": {
              "$ref": "#/components/schemas/PresetDescriptionModel"
            },
            "type": "array",
            "title": "Presets"
          },
          "selectable_record_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selectable Record Ids"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "revision",
          "name",
          "description",
          "stages",
          "presets",
          "selectable_record_ids"
        ],
        "title": "TaskGraphDescriptionModel",
        "description": "Backend-owned, transport-safe description of one task."
      },
...
    }
  }
}